+++ .next/server/chunks/7889e_mimosaweb__next-internal_server_app_api_mindbody_test_route_actions_53899ef2.js
@@ -1,0 +1,3 @@
+module.exports=[52878,(e,o,d)=>{}];
+
+//# sourceMappingURL=7889e_mimosaweb__next-internal_server_app_api_mindbody_test_route_actions_53899ef2.js.map