+++ .next/server/chunks/80b94_GitHub_mimosaweb__next-internal_server_app_api_health_route_actions_926f1933.js
@@ -1,0 +1,3 @@
+module.exports=[26902,(e,o,d)=>{}];
+
+//# sourceMappingURL=80b94_GitHub_mimosaweb__next-internal_server_app_api_health_route_actions_926f1933.js.map